Extracted manipulation of viz.json and vis.instantiateMap to create-vis
