+++ src/vis/vis.js
@@ -24,7 +24,6 @@ var WindshaftClient = require('../windshaft/client');
 var WindshaftLayerGroupConfig = require('../windshaft/layergroup-config');
 var WindshaftNamedMapConfig = require('../windshaft/namedmap-config');
 var WindshaftMap = require('../windshaft/windshaft-map');
-var VizJSON = require('./vizjson');
 
 /**
  * Visualization creation
@@ -189,11 +188,8 @@ var Vis = View.extend({
     });
   },
 
-  load: function (vizjsonData, options) {
+  load: function (vizjson, options) {
     options = options || {};
-
-    var vizjson = new VizJSON(vizjsonData);
-    this._applyOptionsToVizJSON(vizjson, options);
 
     this._dataviewsCollection = new DataviewCollection();
 
@@ -306,7 +302,7 @@ var Vis = View.extend({
 
     // Create the Layer Models and set them on hte map
     this.https = (window && window.location.protocol && window.location.protocol === 'https:') || !!vizjson.https || !!options.https;
-    var layerModels = this._newLayerModels(vizjsonData, this.map);
+    var layerModels = this._newLayerModels(vizjson, this.map);
 
     var infowindowManager = new InfowindowManager(this);
     infowindowManager.manage(this.mapView, this.map);
@@ -317,7 +313,7 @@ var Vis = View.extend({
     // Create the collection of Overlays
     var overlaysCollection = new Backbone.Collection();
     overlaysCollection.bind('reset', function (overlays) {
-      this._addOverlays(overlays, vizjsonData, options);
+      this._addOverlays(overlays, vizjson, options);
     }, this);
     overlaysCollection.reset(vizjson.overlays);
 
@@ -329,10 +325,6 @@ var Vis = View.extend({
       windshaftMap: windshaftMap
     });
 
-    if (!options.skipMapInstantiation) {
-      this.instantiateMap();
-    }
-
     // Lastly: reset the layer models on the map
     this.map.layers.reset(layerModels);
 
@@ -340,67 +332,6 @@ var Vis = View.extend({
     window.vis = this;
 
     return this;
-  },
-
-  _applyOptionsToVizJSON: function (vizjson, options) {
-    vizjson.scrollwheel = options.scrollwheel || vizjson.scrollwheel;
-
-    if (!options.tiles_loader || !options.loaderControl) {
-      vizjson.removeLoaderOverlay();
-    }
-
-    if (options.searchControl === true) {
-      vizjson.addSearchOverlay();
-    } else if (options.searchControl === false) {
-      vizjson.removeSearchOverlay();
-    }
-
-    if ((options.title && vizjson.title) || (options.description && vizjson.description)) {
-      vizjson.addHeaderOverlay(options.title, options.description, options.shareable);
-    }
-
-    if (options.layer_selector) {
-      vizjson.addLayerSelectorOverlay();
-    }
-
-    if (options.zoomControl !== undefined && !options.zoomControl) {
-      vizjson.removeZoomOverlay();
-    }
-
-    // if bounds are present zoom and center will not taken into account
-    var zoom = parseInt(options.zoom, 10);
-    if (!isNaN(zoom)) {
-      vizjson.setZoom(zoom);
-    }
-
-    // Center coordinates?
-    var center_lat = parseFloat(options.center_lat);
-    var center_lon = parseFloat(options.center_lon);
-    if (!isNaN(center_lat) && !isNaN(center_lon)) {
-      vizjson.setCenter([center_lat, center_lon]);
-    }
-
-    // Center object
-    if (options.center !== undefined) {
-      vizjson.setCenter(options.center);
-    }
-
-    // Bounds?
-    var sw_lat = parseFloat(options.sw_lat);
-    var sw_lon = parseFloat(options.sw_lon);
-    var ne_lat = parseFloat(options.ne_lat);
-    var ne_lon = parseFloat(options.ne_lon);
-
-    if (!isNaN(sw_lat) && !isNaN(sw_lon) && !isNaN(ne_lat) && !isNaN(ne_lon)) {
-      vizjson.setBounds([
-        [ sw_lat, sw_lon ],
-        [ ne_lat, ne_lon ]
-      ]);
-    }
-
-    if (options.gmaps_base_type) {
-      vizjson.enforceGMapsBaseLayer(options.gmaps_base_type, options.gmaps_style);
-    }
   },
 
   /**
